+++ lib/mockery.ex
@@ -89,6 +89,25 @@ defmodule Mockery do
   defp to_mod(mod) when is_atom(mod), do: mod
   defp to_mod(mod) when is_binary(mod), do: Module.concat([mod])
 
+  @typedoc "Function name (an atom identifying the function)."
+  @type function_name :: atom()
+
+  @typedoc """
+  A static mock return value
+
+  Any Elixir term except functions.
+  """
+  @type static_mock_value :: term()
+
+  @typedoc """
+  A dynamic mock function.
+
+  This is a function invoked when the mocked function is called.
+  The dynamic mock must accept the same arity as the original function
+  being mocked and its return value is used as the mock result.
+  """
+  @type dynamic_mock :: fun()
+
   @doc """
   Function used to create mock in context of single test process.
 
@@ -125,6 +144,13 @@ defmodule Mockery do
       |> mock([fun2: 1], &string/1)
 
   """
+  # because older version had issue with overlapping @spec
+  if Version.match?(System.version(), "~> 1.17") do
+    @spec mock(module(), function_name(), static_mock_value()) :: module()
+    @spec mock(module(), [{function_name(), arity()}], static_mock_value()) :: module()
+    @spec mock(module(), [{function_name(), arity()}], dynamic_mock()) :: module()
+  end
+
   def mock(mod, fun, value \\ :mocked)
 
   def mock(mod, fun, value) when is_atom(fun) and is_function(value) do
@@ -138,9 +164,34 @@ defmodule Mockery do
     """
   end
 
-  def mock(mod, fun, nil), do: do_mock(mod, fun, Mockery.Nil)
-  def mock(mod, fun, false), do: do_mock(mod, fun, Mockery.False)
-  def mock(mod, fun, value), do: do_mock(mod, fun, value)
+  def mock(mod, [{fun, arity}], value)
+      when is_atom(fun) and is_integer(arity) and is_function(value, arity) do
+    do_mock(mod, [{fun, arity}], value)
+  end
+
+  def mock(_mod, [{fun, arity}], value)
+      when is_atom(fun) and is_integer(arity) and is_function(value) do
+    {:arity, mock_arity} = :erlang.fun_info(value, :arity)
+
+    raise Mockery.Error, """
+    Dynamic mock must have the same arity as the original function
+
+    Original arity: #{arity}
+    Mock arity: #{mock_arity}
+    """
+  end
+
+  def mock(mod, [{fun, arity}], value)
+      when is_atom(fun) and is_integer(arity) and not is_function(value) do
+    do_mock(mod, [{fun, arity}], value)
+  end
+
+  def mock(mod, fun, value) when is_atom(fun) do
+    do_mock(mod, fun, value)
+  end
+
+  defp do_mock(mod, fun, nil), do: do_mock(mod, fun, Mockery.Nil)
+  defp do_mock(mod, fun, false), do: do_mock(mod, fun, Mockery.False)
 
   defp do_mock(mod, fun, value) do
     Utils.put_mock(mod, fun, value)